+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 - Settings: Move Experimental Features into Advanced
 - Settings: Move Connectivity button down and give it an icon (like on android and ios)
 - update in-app help (2023-06-07)
-
+- If clipboard contains a file and text, then only paste the file into deltachat. #3261
 
 <a id="1_37_0"></a>